backend/requirements.txt
fastAPI
supabase
sqlalchemy
uvicorn
psycopg2
python-dotenv
requests